Renamed package to kubeless and corrected Functions and Trigger URI by adding trailing slash
