+++ clients/kubeless-client/src/main/java/com/hashmapinc/kubeless/models/V1beta1AbstractType.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hashmapinc.models;
+package com.hashmapinc.kubeless.models;
 
 import com.google.gson.annotations.SerializedName;
 import io.kubernetes.client.models.V1ObjectMeta;
@@ -23,7 +23,7 @@ import java.util.Objects;
 public abstract class V1beta1AbstractType<T extends Spec> {
 
     @SerializedName("apiVersion")
-    private String apiVersion = null;
+    private String apiVersion = "kubeless.io/v1beta1";
 
     @SerializedName("metadata")
     private V1ObjectMeta metadata = null;
