+++ clients/kubeless-client/src/main/java/com/hashmapinc/kubeless/models/triggers/V1beta1CronJobTrigger.java
@@ -13,17 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hashmapinc.models.triggers;
+package com.hashmapinc.kubeless.models.triggers;
 
 import com.google.gson.annotations.SerializedName;
-import com.hashmapinc.models.V1beta1AbstractTypeList;
+import com.hashmapinc.kubeless.models.V1beta1AbstractType;
 
-public class V1beta1HttpTriggerList extends V1beta1AbstractTypeList<V1beta1HttpTrigger> {
+public class V1beta1CronJobTrigger extends V1beta1AbstractType<V1beta1HttpTriggerSpec>{
 
     @SerializedName("kind")
-    protected String kind = "HTTPTriggerList";
+    protected String kind = "CronJobTrigger";
 
-    public V1beta1HttpTriggerList kind(String kind) {
+    public V1beta1CronJobTrigger kind(String kind) {
         this.kind = kind;
         return this;
     }
